Cleaned NetworkUtil to incorporate new Refactoring in API. Added JSON Wrapping and UnWrapping Functionality. much better design.

+++ app/src/main/java/com/example/savaari_driver/Util.java
@@ -103,6 +103,6 @@ public class Util extends AppCompatActivity
 
         int meterConversion = 1609;
 
-        return (double) (distance * meterConversion);
+        return distance * meterConversion;
     }
 }
